+++ logenc/process.go
@@ -334,226 +334,3 @@ func FileMD5(path string) string {
 	}
 	return fmt.Sprintf("%x", h.Sum(nil))
 }
-
-func Createdir(path string) {
-	fileN := filepath.Base(path)
-	//Create a folder/directory at a full qualified path
-	err := os.Mkdir("./repdata/"+fileN, 0755)
-	if err != nil {
-		log.Fatal(err)
-	}
-}
-func CopyLogs(path string) {
-	fileN := filepath.Base(path)
-	Createdir(path)
-	original, err := os.Open(path)
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer original.Close()
-
-	if CheckFileSum(path) == true {
-		old, err := os.OpenFile("./repdata/"+fileN+"/"+fileN+"old", os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0666)
-
-		if err != nil {
-			log.Fatal(err)
-		}
-		defer old.Close()
-
-		bytesWritten, err := io.Copy(old, original)
-		if err != nil {
-			log.Fatal(err)
-		}
-		fmt.Printf("Bytes Written: %d\n", bytesWritten)
-	} else {
-		new, err := os.OpenFile("./repdata/"+fileN+"/"+fileN+"new", os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0666)
-		if err != nil {
-			log.Fatal(err)
-		}
-		defer new.Close()
-
-		bytesWritten, err := io.Copy(new, original)
-		if err != nil {
-			log.Fatal(err)
-		}
-		fmt.Printf("Bytes Written in new file: %d\n", bytesWritten)
-
-		err = os.Remove("./repdata/" + fileN + "/" + fileN + "new")
-		if err != nil {
-			log.Fatal(err)
-		}
-		err = os.Remove("./repdata/" + fileN + "/" + fileN + "old")
-		if err != nil {
-			log.Fatal(err)
-		}
-
-	}
-
-}
-
-/*
-func Comparefiles(path1 string, path2 string, path string) {
-
-	var data1 LogList
-	var data2 LogList
-	//var wg sync.WaitGroup
-	var str1 string
-	var str2 string
-
-	fileN := path
-	var wait int32 = 0
-	new, err := os.OpenFile("/home/nik/projects/Course/logi2/repdata/Test/"+fileN, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0666)
-	if err != nil {
-		log.Fatal(err)
-	}
-	//	ch := make(chan string, 100)
-
-		for i := runtime.NumCPU() + 1; i > 0; i-- {
-			go func() {
-				wg.Add(1)
-				defer wg.Done()
-				for {
-					if wait == 0 || wait == -1 {
-						line1 = <-ch
-					}
-
-					if wait == 0 || wait == 1 {
-						line2 = <-ch
-					}
-
-	for {
-		err1 := ReadLines(path1, func(line1 string) {
-			if wait == 0 || wait == -1 {
-				str1 = line1
-			}
-		})
-		if err1 != nil {
-			log.Fatalf("ReadLines: %s", err1)
-		}
-
-		err2 := ReadLines(path2, func(line2 string) {
-			if wait == 0 || wait == 1 {
-				str2 = line2
-			}
-		})
-		if err2 != nil {
-			log.Fatalf("ReadLines: %s", err2)
-		}
-
-		data1 = ProcLineDecodeXML(str1)
-		ulid1, _ := ulid.ParseStrict(data1.XML_RECORD_ROOT[0].XML_ULID)
-		data2 = ProcLineDecodeXML(str2)
-		ulid2, _ := ulid.ParseStrict(data2.XML_RECORD_ROOT[0].XML_ULID)
-
-		if ulid1.Compare(ulid2) == 1 {
-			new.Write([]byte(str2))
-			//atomic.AddInt32(&counter, 1)
-			atomic.StoreInt32(&wait, 1)
-
-		} else if ulid1.Compare(ulid2) == -1 {
-			new.Write([]byte(str1))
-			atomic.StoreInt32(&wait, -1)
-
-		} else {
-			new.Write([]byte(str1))
-		}
-
-	}
-
-	//}()
-	//	}
-	/*
-		err1 := ReadLines(path1, func(line1 string) {
-			ch <- line1
-		})
-		err2 := ReadLines(path2, func(line2 string) {
-			ch <- line2
-		})
-		if err1 != nil {
-			log.Fatalf("ReadLines: %s", err1)
-		}
-		if err2 != nil {
-			log.Fatalf("ReadLines: %s", err2)
-		}
-		close(ch)
-		wg.Wait()
-
-*/
-/*
-func Comparefiles2(path1 string, path2 string, path string) {
-
-	//var str1 LogList
-	//	var str2 LogList
-	ch1 := make(chan string, 100)
-
-	ch2 := make(chan string, 100)
-
-	//fileN := path
-	//var wait int32 = 0
-	new, err := os.OpenFile(path, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0666)
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer new.Close()
-	file1, err := os.OpenFile(path1, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0666)
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer file1.Close()
-
-	file2, err := os.OpenFile(path2, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0666)
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer file2.Close()
-
-	scanner1 := bufio.NewScanner(file1)
-	scanner2 := bufio.NewScanner(file2)
-
-	//info1, err := os.Stat(path1)
-	//info2, err := os.Stat(path2)
-
-	go func() {
-		for {
-			select {
-			case line1, _ := <-ch1:
-				ulid1, _ := ulid.ParseStrict(line1.XML_RECORD_ROOT[0].XML_ULID)
-
-			case line2, _ := <-ch2:
-				ulid2, _ := ulid.ParseStrict(line2.XML_RECORD_ROOT[0].XML_ULID)
-
-			}
-		}
-
-	}()
-
-	for scanner1.Scan() {
-
-		ch1 <- ProcLine(scanner1.Text())
-
-	}
-	for scanner1.Scan() {
-		ch2 <- ProcLine(scanner2.Text())
-
-	}
-	/*
-		for scanner1.Scan() || scanner2.Scan() {
-			str1 = ProcLineDecodeXML(scanner1.Text())
-
-			str2 = ProcLineDecodeXML(scanner2.Text())
-
-			ulid1, _ := ulid.ParseStrict(str1.XML_RECORD_ROOT[0].XML_ULID)
-			ulid2, _ := ulid.ParseStrict(str2.XML_RECORD_ROOT[0].XML_ULID)
-
-			if ulid1.Compare(ulid2) == 1 {
-				new.Write([]byte(scanner2.Text()))
-
-			} else {
-				new.Write([]byte(scanner1.Text()))
-
-			}
-		}
-
-
-}
-*/
