--- logenc/test0
@@ -1,7 +1,0 @@
-7TMCS TEST,3/TEST/TEST,290,2,Sunday 10 Oct 2021,0,01FM2G2ARAX54VX7DJE0BZT48NСостояние '203.76.243.197Cервер КС_UDP/Пинг',Context:  -- void tmcs::AbstractMonitor::,Clackmannanshire
-7TMCS TEST,3/TEST/TEST,290,2,Wednesday 25 Aug 2021,0,01FM2G2BX6G9ARW8ERWSP6E6K3Состояние '52.103.226.180Cервер КС_UDP/Пинг',Context:  -- void tmcs::AbstractMonitor::,West Lothian
-7TMCS TEST,3/TEST/TEST,290,2,Monday 12 Jul 2021,0,01FM2G2D6WM23EPFJ3W2HATZD3Состояние '234.46.117.60Cервер КС_UDP/Пинг',Context:  -- void tmcs::AbstractMonitor::,Rutland
-7TMCS TEST,3/TEST/TEST,290,2,Friday 15 Jan 2021,0,01FM2G2EERZHVD6B3D4SMHNGMSСостояние '63.188.65.210Cервер КС_UDP/Пинг',Context:  -- void tmcs::AbstractMonitor::,Cornwall
-7TMCS TEST,3/TEST/TEST,290,2,Tuesday 2 Nov 2021,0,01FM2G2FT8MVXP8G4XZW985R89Состояние '161.187.180.141Cервер КС_UDP/Пинг',Context:  -- void tmcs::AbstractMonitor::,Essex
-7TMCS TEST,3/TEST/TEST,290,2,Monday 13 Dec 2021,0,01FM2G2H879RCMBN3RPQP2JE39Состояние '66.125.54.231Cервер КС_UDP/Пинг',Context:  -- void tmcs::AbstractMonitor::,Midlothian
-7TMCS TEST,3/TEST/TEST,290,2,Tuesday 23 Nov 2021,0,01FM7510X1HWDBXN4G5ZC5BM1MСостояние '50.115.79.253Cервер КС_UDP/Пинг',Context:  -- void tmcs::AbstractMonitor::,Carmarthenshire
