--- logenc/test1
@@ -1,7 +1,0 @@
-2TMCS TEST,3/TEST/TEST,821,4,Friday 26 Nov 2021,0,01FM2G2ARAT07S2V80F0F021S4Состояние '28.78.44.206Cервер КС_UDP/Пинг',Context:  -- void tmcs::AbstractMonitor::,Wiltshire
-2TMCS TEST,3/TEST/TEST,821,4,Monday 23 Aug 2021,0,01FM2G2BX6535Z5BTY468BBHW6Состояние '196.8.2.203Cервер КС_UDP/Пинг',Context:  -- void tmcs::AbstractMonitor::,Cumbria
-2TMCS TEST,3/TEST/TEST,821,4,Tuesday 7 Dec 2021,0,01FM2G2D6W45G2J8P3ZZ48EB6FСостояние '144.240.74.30Cервер КС_UDP/Пинг',Context:  -- void tmcs::AbstractMonitor::,Down
-2TMCS TEST,3/TEST/TEST,821,4,Monday 20 Sep 2021,0,01FM2G2EER27C574X0WFB5GH7KСостояние '148.201.203.100Cервер КС_UDP/Пинг',Context:  -- void tmcs::AbstractMonitor::,Down
-2TMCS TEST,3/TEST/TEST,821,4,Monday 5 Apr 2021,0,01FM2G2FT8QFWBNDEW00KYE91YСостояние '247.89.148.130Cервер КС_UDP/Пинг',Context:  -- void tmcs::AbstractMonitor::,Warwickshire
-2TMCS TEST,3/TEST/TEST,821,4,Monday 1 Feb 2021,0,01FM2G2H87WQY2678C479QBCA8Состояние '169.206.45.233Cервер КС_UDP/Пинг',Context:  -- void tmcs::AbstractMonitor::,West Yorkshire
-2TMCS TEST,3/TEST/TEST,821,4,Sunday 14 Mar 2021,0,01FM7510X11KR3RKM854G7YG3EСостояние '252.2.176.122Cервер КС_UDP/Пинг',Context:  -- void tmcs::AbstractMonitor::,Cheshire
